refactor(viztooltip): update import paths and names where used
