+++ public/app/features/explore/flotgraph/types.ts
@@ -1,4 +1,4 @@
-import { VizTooltipOptions, LegendDisplayMode, LegendPlacement } from '@grafana/ui';
+import { VizTooltipOptions, TooltipDisplayMode, LegendDisplayMode, LegendPlacement } from '@grafana/ui';
 import { YAxis } from '@grafana/data';
 
 export interface SeriesOptions {
@@ -35,7 +35,7 @@ export const defaults: Options = {
     placement: 'bottom',
   },
   series: {},
-  tooltipOptions: { mode: 'single' },
+  tooltipOptions: { mode: TooltipDisplayMode.Single },
 };
 
 export interface GraphLegendEditorLegendOptions {
