+++ public/app/plugins/panel/timeseries/plugins/AnnotationMarker.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback, useRef, useState } from 'react';
 import { GrafanaTheme } from '@grafana/data';
-import { HorizontalGroup, Portal, Tag, TooltipContainer, useStyles } from '@grafana/ui';
+import { HorizontalGroup, Portal, Tag, VizTooltipContainer, useStyles } from '@grafana/ui';
 import { css } from '@emotion/css';
 
 interface AnnotationMarkerProps {
@@ -38,7 +38,7 @@ export const AnnotationMarker: React.FC<AnnotationMarkerProps> = ({ time, text, 
     const elBBox = el.getBoundingClientRect();
 
     return (
-      <TooltipContainer
+      <VizTooltipContainer
         position={{ x: elBBox.left, y: elBBox.top + elBBox.height }}
         offset={{ x: 0, y: 0 }}
         onMouseEnter={onMouseEnter}
@@ -61,7 +61,7 @@ export const AnnotationMarker: React.FC<AnnotationMarkerProps> = ({ time, text, 
             </>
           </div>
         </div>
-      </TooltipContainer>
+      </VizTooltipContainer>
     );
   }, [onMouseEnter, onMouseLeave, styles, time, text, tags]);
